Changed modis data to fit multi dir loading
